[DEV-593] fixing wrong imports due to moving blockchain enum and other merge conflicts left overs
