+++ src/payment/models/buy/__tests__/buy.controller.spec.ts
@@ -10,6 +10,7 @@ import { BuyCryptoService } from '../../buy-crypto/services/buy-crypto.service';
 import { createDefaultBuy } from './mock/buy.entity.mock';
 import { UserRole } from 'src/shared/auth/user-role.enum';
 import { TestUtil } from 'src/shared/test.util';
+import { Blockchain } from 'src/ain/services/crypto.service';
 
 describe('BuyController', () => {
   let controller: BuyController;
@@ -50,7 +51,7 @@ describe('BuyController', () => {
   it('should return a min deposit of 1 for a default buy route', async () => {
     jest.spyOn(buyService, 'getUserBuys').mockResolvedValue([createDefaultBuy()]);
 
-    await expect(controller.getAllBuy({ id: 0, address: '', role: UserRole.USER })).resolves.toMatchObject([
+    await expect(controller.getAllBuy({ id: 0, address: '', role: UserRole.USER, blockchain: Blockchain.DEFICHAIN })).resolves.toMatchObject([
       { minDeposits: [{ amount: 1, asset: 'USD' }] },
     ]);
   });
