+++ src/payment/models/crypto-route/__tests__/crypto-route.controller.spec.ts
@@ -10,6 +10,7 @@ import { BuyCryptoService } from '../../buy-crypto/services/buy-crypto.service';
 import { TestUtil } from 'src/shared/test.util';
 import { UserRole } from 'src/shared/auth/user-role.enum';
 import { createDefaultCryptoRoute } from './mock/crypto-route.entity.mock';
+import { Blockchain } from 'src/ain/services/crypto.service';
 
 describe('CryptoRouteController', () => {
   let controller: CryptoRouteController;
@@ -50,7 +51,7 @@ describe('CryptoRouteController', () => {
   it('should return a min deposit of 0.0005 for a default crypto route', async () => {
     jest.spyOn(cryptoRouteService, 'getUserCryptos').mockResolvedValue([createDefaultCryptoRoute()]);
 
-    await expect(controller.getAllCrypto({ id: 0, address: '', role: UserRole.USER })).resolves.toMatchObject([
+    await expect(controller.getAllCrypto({ id: 0, address: '', role: UserRole.USER, blockchain: Blockchain.DEFICHAIN })).resolves.toMatchObject([
       { minDeposits: [{ amount: 0.0005, asset: 'BTC' }] },
     ]);
   });
