+++ src/payment/models/crypto-route/__tests__/mock/crypto-route.entity.mock.ts
@@ -1,4 +1,4 @@
-import { Blockchain } from 'src/ain/node/node.service';
+import { Blockchain } from 'src/ain/services/crypto.service';
 import {
   createCustomDeposit,
   createDefaultDeposit,
